requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database (for future PostgreSQL integration)
sqlalchemy==2.0.23

# Type checking and development
mypy==1.7.1
typing-extensions==4.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
